Append created div if no element was passed
It turns out, in fact, that the previous version of this component allowed
itself to be used without any template at all. Sure, the javascript didn't work
and it wasn't documented at all but we don't know that someone isn't using that
feature and seeing as this is a legal requirement it's probably best to not
change its behaviour during a crunchy time.

Of course, legally speaking, this component is dubious anyway because even if
you opt out of tracking cookies The Financial Times continues to track you
anyway, they just stop showing you personalised ads.