Remove the standard theme option (move to base styles).
The following Sass from the README did not work:
```scss
@include oCookieMessage($opts: (
	'themes': ('alternative')
));
```

That's because the `alternative` theme builds on the `standard`
theme, with no modifier class `o-message--standard` the `standard`
theme effects all themes.

This commit removes `standard` as a theme and moves its styles
within the `_oCookieMessageBase` mixin, to be output by default
for other themes to build from.